requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
structlog
orjson
pytest
pytest-asyncio
httpx
black
isort
ruff
mypy
pre-commit
supabase
stripe
python-jose[cryptography]
pyyaml
